3.7

+++ src/main/java/calebzhou/rdi/microservice/ctrler/v37/Island2Ctrler.java
@@ -1,14 +1,16 @@
 package calebzhou.rdi.microservice.ctrler.v37;
 
+import calebzhou.rdi.microservice.annotation.PidTokenCheck;
 import calebzhou.rdi.microservice.constant.ResultCode;
 import calebzhou.rdi.microservice.dao.Island2Mapper;
+import calebzhou.rdi.microservice.utils.JwtUtils;
 import calebzhou.rdi.microservice.utils.RdiSerializer;
 import calebzhou.rdi.microservice.model.Island2;
 import calebzhou.rdi.microservice.utils.TimeUtils;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/v37/island2")
+@RequestMapping("/v37/island")
 public class Island2Ctrler {
     final Island2Mapper mapper;
     public Island2Ctrler(Island2Mapper mapper) {
@@ -16,8 +18,9 @@ public class Island2Ctrler {
     }
 
     //获取岛屿基本信息 成功返回岛屿数据
+    @PidTokenCheck
     @RequestMapping(value = "/{pid}",method = RequestMethod.GET)
-    public Object getIsland(@PathVariable String pid){
+    public Object getIsland(@RequestHeader("rauth") String token,@PathVariable String pid){
         boolean playerOwnIsland = mapper.isPlayerOwnIsland(pid);
         boolean playerJoinAnyIsland = mapper.isPlayerJoinAnyIsland(pid);
         if(!playerOwnIsland && !playerJoinAnyIsland){
@@ -36,8 +39,9 @@ public class Island2Ctrler {
     }
 
     //提供玩家pid创建新的空岛，成功返回岛屿id
+    @PidTokenCheck
     @RequestMapping(value = "/{pid}",method = RequestMethod.POST)
-    public Object createIsland(@PathVariable String pid){
+    public Object createIsland(@RequestHeader("rauth") String token,@PathVariable String pid){
         if(mapper.isPlayerOwnIsland(pid)){
             return ResultCode.sourceAlreadyOwnIsland;
         }
@@ -51,8 +55,9 @@ public class Island2Ctrler {
         return mapper.getPlayerOwnIslandId(pid);
     }
     //提供玩家pid删除空岛，返回删除的岛屿id
+    @PidTokenCheck
     @RequestMapping(value = "/{pid}",method = RequestMethod.DELETE)
-    public Object deleteIsland(@PathVariable String pid){
+    public Object deleteIsland(@RequestHeader("rauth") String token,@PathVariable String pid){
         //没岛不删
         if(!mapper.isPlayerOwnIsland(pid)){
             return ResultCode.sourceNotOwnIsland;
@@ -64,8 +69,9 @@ public class Island2Ctrler {
         return iid;
     }
     //提供玩家pid修改岛主
+    @PidTokenCheck
     @RequestMapping(value = "/transfer/{pid}/{targetPid}",method = RequestMethod.PUT)
-    public ResultCode transferIsland(@PathVariable String pid,@PathVariable String targetPid){
+    public ResultCode transferIsland(@RequestHeader("rauth") String token,@PathVariable String pid,@PathVariable String targetPid){
         //不能转给自己,自己必须有岛，自己不能加岛，对方不能有岛，对方不能加岛，
         if(pid.equals(targetPid)){
             return ResultCode.sourceEqualsTarget;
@@ -88,8 +94,9 @@ public class Island2Ctrler {
     }
 
     //提供玩家pid修改空岛传送点，参数x,y,z,w,p坐标 0失败1成功
+    @PidTokenCheck
     @RequestMapping(value = "/loca/{pid}",method = RequestMethod.PUT)
-    public ResultCode changeLocation(@PathVariable String pid,@RequestParam String dim, @RequestParam double x,@RequestParam double y,@RequestParam double z,
+    public ResultCode changeLocation(@RequestHeader("rauth") String token,@PathVariable String pid,@RequestParam String dim, @RequestParam double x,@RequestParam double y,@RequestParam double z,
                               @RequestParam double w,@RequestParam double p){
         //没有岛屿改不了
         if(!mapper.isPlayerOwnIsland(pid)){
@@ -111,8 +118,9 @@ public class Island2Ctrler {
     }
 
     //添加空岛成员，提供岛主pid和成员pid,1成功0失败
+    @PidTokenCheck
     @RequestMapping(value = "/crew/{pid}/{mpid}",method = RequestMethod.POST)
-    public ResultCode addMember(@PathVariable String pid,@PathVariable String mpid){
+    public ResultCode addMember(@RequestHeader("rauth") String token,@PathVariable String pid,@PathVariable String mpid){
         //没岛不加
         if(!mapper.isPlayerOwnIsland(pid))
             return ResultCode.sourceNotOwnIsland;
@@ -132,14 +140,25 @@ public class Island2Ctrler {
         return ResultCode.success;
     }
     //删除空岛成员，提供岛主pid和成员pid 1成功0失败
-    @RequestMapping(value = "/crew/{mpid}",method = RequestMethod.DELETE)
-    public ResultCode removeMember(@PathVariable String mpid){
+    @PidTokenCheck
+    @RequestMapping(value = "/crew/{pid}/{mpid}",method = RequestMethod.DELETE)
+    public ResultCode removeMember(@RequestHeader("rauth") String token,@PathVariable String pid,@PathVariable String mpid){
+        //自己没有岛 不删
+        if(!mapper.isPlayerOwnIsland(pid))
+            return ResultCode.sourceNotOwnIsland;
+        //不能删除自己
+        if(pid.equals(mpid))
+            return ResultCode.sourceEqualsTarget;
         //目标有自己的岛，不删
         if(mapper.isPlayerOwnIsland(mpid))
             return ResultCode.targetAlreadyOwnIsland;
         //目标没进任何岛 不删
         if(!mapper.isPlayerJoinAnyIsland(mpid))
             return ResultCode.targetNotJoinAnyIsland;
+        Integer iidOwn = mapper.getPlayerOwnIslandId(pid);
+        //目标加入的岛不归自己 不删
+        if(mapper.isPlayerJoinAnyIsland(mpid) && !mapper.isPlayerJoinOneIsland(mpid,iidOwn))
+            return ResultCode.sourceNotOwnTarget;
         Integer iid = mapper.getPlayerJoinIslandId(mpid);
         mapper.deleteMember(mpid,iid);
         return ResultCode.success;
